Fixed lexer for rust parser
How lexer works: it takes lex rules as regexps and tries match each with string, and it assumes each possible matching will be at begging of string being scanned:
for this it prepends ^ char for each regular expression. But, it doesn't work for multiline regepxs.

So, this commit switches regexps back to singleline mode for rust plugin.